Create a dedicated AWS organization preset (#1335)
